--- golang/internal/sample.go
@@ -1,5 +1,0 @@
-package internal
-
-func Summ(a, b int) int {
-	return a + b
-}
